ongoing restructure
